requirements/base.txt
amqp==2.2.1
babel==2.5.1
billiard==3.5.0.3
celery==4.1.0
contextlib2==0.5.5
coverage==4.4.1
cython==0.26
django==1.11.3
django-allauth==0.32.0
django-autoslug==1.9.3
django-braces==1.11.0
django-crispy-forms==1.6.1
django-environ==0.4.3
django-extensions==1.8.1
django-filter==1.0.4
django-floppyforms==1.7.0
django-model-utils==3.0.0
django-phonenumber-field==1.3.0
django-redis==4.8.0
django-secure==1.0.1
django-webpack-loader==0.5
djangorestframework==3.6.3
factory-boy==2.9.2
Faker==0.7.18
fitbit==0.3.0
ipython==4.2
kombu==4.1.0
Markdown==2.6.8
numpy==1.13.1
oauthlib==2.0.2
openpyxl==2.4.8
pandas==0.20.3
patsy==0.4.1
phonenumberslite==8.8.2
Pillow==3.2.0
psycopg2==2.6.1
PyJWT==1.5.3
pysocks==1.6.7
pytz==2017.2
raven==6.1.0
readline==6.2.4.1
redis==2.10.5
requests==2.18.3
requests-oauthlib==0.8.0
scipy==0.19.1
setuptools==36.2.7
simplejson==3.11.1
sqlparse==0.2.3
twilio==6.6.3
unicode-slugify==0.1.3
uuid==1.30
vine==1.1.4
wheel==0.29.0
whitenoise==3.3.0
xlrd==1.0.0
XlsxWriter==0.9.8
